+++ lib/core/widgets/app_fields.dart
@@ -11,6 +11,7 @@ class AppFormField extends StatelessWidget {
   final IconData? prefixIcon;
   final bool readOnly;
   final VoidCallback? onTap;
+  final ValueChanged<String>? onChanged;
 
   const AppFormField({
     super.key,
@@ -22,6 +23,7 @@ class AppFormField extends StatelessWidget {
     this.prefixIcon,
     this.readOnly = false,
     this.onTap,
+    this.onChanged,
   });
 
   @override
@@ -44,6 +46,7 @@ class AppFormField extends StatelessWidget {
           validator: validator,
           readOnly: readOnly,
           onTap: onTap,
+          onChanged: onChanged,
           decoration: InputDecoration(
             hintText: hintText,
             prefixIcon: prefixIcon != null