+++ lib/features/publications/domain/publication_repository.dart
@@ -3,6 +3,7 @@ import 'publication.dart';
 abstract interface class PublicationRepository {
   Future<Publication> create(Publication publication);
   Future<List<Publication>> getAll();
+  Future<List<Publication>> getActivePublications();
   Future<Publication?> getById(int id);
   Future<List<Publication>> searchByName(String query, {int limit = 20});
   Future<List<Publication>> searchByCode(String query, {int limit = 20});